display: carve webrtc offer.rs (pure move)
